--- src/controllers/index.ts
@@ -1,2 +1,0 @@
-export * from './app.controller';
-export * from './profile.controller';
