+++ src/profile/index.ts
@@ -1,0 +1,4 @@
+export * from './dto';
+export * from './profile.controller';
+export * from './profile.service';
+export * from './profile.module';
