--- src/services/index.ts
@@ -1,2 +1,0 @@
-export * from './app.service';
-export * from './profile.service';
